# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NexFlow CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /deals:
    get:
      operationId: listDeals
      tags: [deals]
      summary: List deals
      parameters:
        - name: stage
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of deals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DealListResponse"
    post:
      operationId: createDeal
      tags: [deals]
      summary: Create deal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDealBody"
      responses:
        "201":
          description: Created deal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
  /deals/{id}:
    get:
      operationId: getDeal
      tags: [deals]
      summary: Get deal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
    patch:
      operationId: updateDeal
      tags: [deals]
      summary: Update deal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDealBody"
      responses:
        "200":
          description: Updated deal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
    delete:
      operationId: deleteDeal
      tags: [deals]
      summary: Delete deal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── Activities ──
...
  /scripts:
    get:
      operationId: listScripts
      tags: [scripts]
      summary: List scripts
      responses:
        "200":
          description: Scripts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Script"
    post:
      operationId: createScript
      tags: [scripts]
      summary: Create script
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScriptBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Script"
  /scripts/{id}:
    get:
      operationId: getScript
      tags: [scripts]
      summary: Get script
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Script
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Script"
    patch:
      operationId: updateScript
      tags: [scripts]
      summary: Update script
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateScriptBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Script"
    delete:
      operationId: deleteScript
      tags: [scripts]
      summary: Delete script
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── Notifications ──
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications
      parameters:
        - name: unread_only
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Notification marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
  /dashboard/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent activity feed
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
...
  /dashboard/top-contacts:
    get:
      operationId: getTopContacts
      tags: [dashboard]
      summary: Get top contacts by score
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Top contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"

  # ── Analytics ──
...
components:
  schemas:
...
    Contact:
      type: object
      properties:
        id:
          type: string
        org_id:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        phone:
          type: string
        title:
          type: string
        company_id:
          type: string
        company_name:
          type: string
        lead_score:
          type: number
        status:
          type: string
          enum: [new, active, qualified, unqualified, customer]
        avatar_url:
          type: string
        linkedin_url:
          type: string
        notes:
          type: string
        tags:
          type: array
          items:
            type: string
        created_at:
          type: string
        updated_at:
          type: string
      required:
        - id
        - first_name
        - last_name
        - status
        - created_at
...
    Deal:
      type: object
      properties:
        id:
          type: string
        org_id:
          type: string
        title:
          type: string
        contact_id:
          type: string
        contact_name:
          type: string
        company_id:
          type: string
        company_name:
          type: string
        stage:
          type: string
          enum: [lead, qualified, proposal, negotiation, closed_won, closed_lost]
        value:
          type: integer
          description: Deal value in smallest currency unit
        currency:
          type: string
          default: USD
        probability:
          type: number
          minimum: 0
          maximum: 100
        expected_close_date:
          type: string
        notes:
          type: string
        tags:
          type: array
          items:
            type: string
        created_at:
          type: string
        updated_at:
          type: string
      required:
        - id
        - title
        - stage
        - created_at
    DealListResponse:
      type: object
      properties:
        deals:
          type: array
          items:
            $ref: "#/components/schemas/Deal"
        total:
          type: integer
        total_value:
          type: integer
      required:
        - deals
        - total
        - total_value
    CreateDealBody:
      type: object
      properties:
        title:
          type: string
        contact_id:
          type: string
        company_id:
          type: string
        stage:
          type: string
          enum: [lead, qualified, proposal, negotiation, closed_won, closed_lost]
        value:
          type: integer
        currency:
          type: string
        probability:
          type: number
        expected_close_date:
          type: string
        notes:
          type: string
      required:
        - title
        - stage
    UpdateDealBody:
      type: object
      properties:
        title:
          type: string
        contact_id:
          type: string
        company_id:
          type: string
        stage:
          type: string
          enum: [lead, qualified, proposal, negotiation, closed_won, closed_lost]
        value:
          type: integer
        currency:
          type: string
        probability:
          type: number
        expected_close_date:
          type: string
        notes:
          type: string
    Activity:
      type: object
      properties:
        id:
          type: string
        org_id:
          type: string
        contact_id:
          type: string
        contact_name:
          type: string
        deal_id:
          type: string
        type:
          type: string
          enum: [call, email, whatsapp, meeting, note, task]
        title:
          type: string
        body:
          type: string
        status:
          type: string
          enum: [pending, completed, cancelled]
        scheduled_at:
          type: string
        completed_at:
          type: string
        created_at:
          type: string
      required:
        - id
        - type
        - title
        - status
        - created_at
...
    Script:
      type: object
      properties:
        id:
          type: string
        org_id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [cold_call, follow_up, demo, objection_handling, closing]
        content:
          type: string
        language:
          type: string
          enum: [en, ar, both]
        tags:
          type: array
          items:
            type: string
        created_at:
          type: string
        updated_at:
          type: string
      required:
        - id
        - name
        - type
        - content
        - created_at
    CreateScriptBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [cold_call, follow_up, demo, objection_handling, closing]
        content:
          type: string
        language:
          type: string
          enum: [en, ar, both]
        tags:
          type: array
          items:
            type: string
      required:
        - name
        - type
        - content
    UpdateScriptBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [cold_call, follow_up, demo, objection_handling, closing]
        content:
          type: string
        language:
          type: string
          enum: [en, ar, both]
        tags:
          type: array
          items:
            type: string
    Notification:
      type: object
      properties:
        id:
          type: string
        org_id:
          type: string
        title:
          type: string
        body:
          type: string
        type:
          type: string
          enum: [signal, deal, call, task, system, ai]
        read:
          type: boolean
        entity_id:
          type: string
        entity_type:
          type: string
        created_at:
          type: string
      required:
        - id
        - title
        - type
        - read
        - created_at
...