# lib/api-zod/src/generated/types/updateDealBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NexFlow CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateDealBodyStage } from "./updateDealBodyStage";

export interface UpdateDealBody {
  title?: string;
  contact_id?: string;
  company_id?: string;
  stage?: UpdateDealBodyStage;
  value?: number;
  currency?: string;
  probability?: number;
  expected_close_date?: string;
  notes?: string;
}